+++ nn_dataflow/tests/unit_test/test_scheduling_result.py
@@ -34,21 +34,22 @@ class TestSchedulingResult(unittest.TestCase):
 
     def setUp(self):
 
-        self.dict_loop = OrderedDict([('cost', 1.234),
-                                      ('time', 123.4),
-                                      ('proc_time', 59),
-                                      ('bus_time', 40),
-                                      ('dram_time', 120),
-                                      ('ops', 1234),
-                                      ('access', [[2, 3, 4],
-                                                  [30, 40, 50],
-                                                  [400, 500, 600],
-                                                  [5000, 6000, 7000]]),
-                                     ])
-        self.dict_part = OrderedDict([('cost', 9.876),
-                                      ('total_nhops', [123, 456, 789]),
-                                      ('num_nodes', 4),
-                                     ])
+        self.scheme = OrderedDict([('cost', 9.876 + 1.234),
+                                   ('time', 123.4),
+                                   ('ops', 1234),
+                                   ('num_nodes', 4),
+                                   ('cost_loop', 1.234),
+                                   ('cost_part', 9.876),
+                                   ('proc_time', 59),
+                                   ('bus_time', 40),
+                                   ('dram_time', 120),
+                                   ('access', [[2, 3, 4],
+                                               [30, 40, 50],
+                                               [400, 500, 600],
+                                               [5000, 6000, 7000]]),
+                                   ('total_nhops', [123, 456, 789]),
+                                   ('fetch', [[1, 2, 1], [3, 4, 5]]),
+                                  ])
 
         part = PartitionScheme(order=range(pe.NUM), pdims=[(1, 1)] * pe.NUM)
         self.ofmap_layout = DataLayout(
@@ -59,48 +60,35 @@ class TestSchedulingResult(unittest.TestCase):
 
     def test_valid_args(self):
         ''' Valid arguments. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
-        self.assertIn('ops', result.dict_loop)
-        self.assertIn('total_nhops', result.dict_part)
+        self.assertIn('ops', result.scheme)
+        self.assertIn('total_nhops', result.scheme)
         self.assertEqual(result.ofmap_layout, self.ofmap_layout)
 
-    def test_invalid_dict_loop(self):
-        ''' Invalid dict_loop. '''
+    def test_invalid_scheme(self):
+        ''' Invalid scheme. '''
         with self.assertRaisesRegexp(TypeError,
-                                     'SchedulingResult: .*dict_loop.*'):
-            _ = SchedulingResult(dict_loop={},
-                                 dict_part=self.dict_part,
-                                 ofmap_layout=self.ofmap_layout)
-
-    def test_invalid_dict_part(self):
-        ''' Invalid dict_part. '''
-        with self.assertRaisesRegexp(TypeError,
-                                     'SchedulingResult: .*dict_part.*'):
-            _ = SchedulingResult(dict_loop=self.dict_loop,
-                                 dict_part={},
+                                     'SchedulingResult: .*scheme.*'):
+            _ = SchedulingResult(scheme={},
                                  ofmap_layout=self.ofmap_layout)
 
     def test_invalid_ofmap_layout(self):
         ''' Invalid ofmap_layout. '''
         with self.assertRaisesRegexp(TypeError,
                                      'SchedulingResult: .*ofmap_layout.*'):
-            _ = SchedulingResult(dict_loop=self.dict_loop,
-                                 dict_part=self.dict_part,
+            _ = SchedulingResult(scheme=self.scheme,
                                  ofmap_layout=None)
 
     def test_total_cost(self):
         ''' Accessor total_cost. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertAlmostEqual(result.total_cost, 1.234 + 9.876)
 
     def test_total_time(self):
         ''' Accessor total_time. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertAlmostEqual(result.total_time, 123.4)
 
@@ -109,89 +97,75 @@ class TestSchedulingResult(unittest.TestCase):
 
     def test_total_node_time(self):
         ''' Accessor total_node_time. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertAlmostEqual(result.total_node_time, max(59, 40))
 
-        dict_loop = self.dict_loop
-        dict_loop['bus_time'] = 100
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        scheme = self.scheme
+        scheme['bus_time'] = 100
+        result = SchedulingResult(scheme=scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertAlmostEqual(result.total_node_time, max(59, 100))
 
     def test_total_dram_time(self):
         ''' Accessor total_dram_time. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertAlmostEqual(result.total_dram_time, 120)
 
     def test_total_proc_time(self):
         ''' Accessor total_proc_time. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertAlmostEqual(result.total_proc_time, 59)
 
-        dict_loop = self.dict_loop
-        dict_loop['bus_time'] = 100
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        scheme = self.scheme
+        scheme['bus_time'] = 100
+        result = SchedulingResult(scheme=scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertAlmostEqual(result.total_proc_time, 59)
 
     def test_total_ops(self):
         ''' Accessor total_ops. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertEqual(result.total_ops, 1234)
 
     def test_total_accesses(self):
         ''' Accessor total_cost. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertSequenceEqual(result.total_accesses,
                                  [9, 120, 1500, 18000])
 
     def test_total_noc_hops(self):
         ''' Accessor total_noc_hops. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertEqual(result.total_noc_hops, 1368)
 
     def test_num_nodes(self):
         ''' Accessor num_nodes. '''
-        result = SchedulingResult(dict_loop=self.dict_loop,
-                                  dict_part=self.dict_part,
+        result = SchedulingResult(scheme=self.scheme,
                                   ofmap_layout=self.ofmap_layout)
         self.assertEqual(result.num_nodes, 4)
 
     def test_cmp_key(self):
         ''' Get cmp_key. '''
-        result1 = SchedulingResult(dict_loop=self.dict_loop,
-                                   dict_part=self.dict_part,
+        result1 = SchedulingResult(scheme=self.scheme,
                                    ofmap_layout=self.ofmap_layout)
 
-        dict_loop = self.dict_loop.copy()
-        dict_loop['cost'] = 2
-        dict_loop['time'] = 12.34
-        dict_part = self.dict_part.copy()
-        dict_part['cost'] = 10
-        result2 = SchedulingResult(dict_loop=dict_loop,
-                                   dict_part=dict_part,
+        scheme = self.scheme.copy()
+        scheme['cost'] = 2 + 10
+        scheme['time'] = 12.34
+        result2 = SchedulingResult(scheme=scheme,
                                    ofmap_layout=self.ofmap_layout)
 
         self.assertGreater(result2.cmp_key(), result1.cmp_key())
 
-        dict_loop = dict_loop.copy()
-        dict_loop['time'] = 23.4
-        result3 = SchedulingResult(dict_loop=dict_loop,
-                                   dict_part=dict_part,
+        scheme = scheme.copy()
+        scheme['time'] = 23.4
+        result3 = SchedulingResult(scheme=scheme,
                                    ofmap_layout=self.ofmap_layout)
 
         self.assertGreater(result3.cmp_key(), result2.cmp_key())